+++ app/Models/Ingredient.php
@@ -4,29 +4,7 @@ namespace App\Models;
 
 use Illuminate\Database\Eloquent\Model;
 
-final /**
- * App\Models\Ingredient
- *
- * @property int $id
- * @property int $recipe_id
- * @property string $text
- * @property string|null $amount
- * @property string|null $unit
- * @property string|null $header
- * @property \Illuminate\Support\Carbon $updated_at
- * @property-read \App\Models\Recipe $recipe
- * @method static \Illuminate\Database\Eloquent\Builder|\App\Models\Ingredient newModelQuery()
- * @method static \Illuminate\Database\Eloquent\Builder|\App\Models\Ingredient newQuery()
- * @method static \Illuminate\Database\Eloquent\Builder|\App\Models\Ingredient query()
- * @method static \Illuminate\Database\Eloquent\Builder|\App\Models\Ingredient whereAmount($value)
- * @method static \Illuminate\Database\Eloquent\Builder|\App\Models\Ingredient whereHeader($value)
- * @method static \Illuminate\Database\Eloquent\Builder|\App\Models\Ingredient whereId($value)
- * @method static \Illuminate\Database\Eloquent\Builder|\App\Models\Ingredient whereRecipeId($value)
- * @method static \Illuminate\Database\Eloquent\Builder|\App\Models\Ingredient whereText($value)
- * @method static \Illuminate\Database\Eloquent\Builder|\App\Models\Ingredient whereUnit($value)
- * @method static \Illuminate\Database\Eloquent\Builder|\App\Models\Ingredient whereUpdatedAt($value)
- * @mixin \Eloquent
- */ class Ingredient extends Model
+class Ingredient extends Model
 {
     protected $table = 'ingredients';
     public $timestamps = false;
