+++ app/Scopes/VisibilityScope.php
@@ -2,7 +2,7 @@
 
 namespace App\Scopes;
 
-use App\User;
+use App\Models\User;
 use Illuminate\Database\Eloquent\Builder;
 use Illuminate\Database\Eloquent\Model;
 use Illuminate\Database\Eloquent\Scope;
